+++ streamdeck-plugin/src/idea-plugin.ts
@@ -7,7 +7,7 @@ import {ActionIdBrowserAction} from "./actions/action-id-browser-action";
 import {DebugAction} from "./actions/debug-action";
 import {EmptyAction} from "./actions/empty-action";
 import {GitPullAction} from './actions/git-pull-action'
-import {NewProjectAction} from './actions/newProject'
+import {NewProjectAction} from './actions/new-project-action'
 import {OpenRecentAction} from "./actions/open-recent-action";
 import {PauseAction} from "./actions/pause-action";
 import {ResumeAction} from "./actions/resume-action";
@@ -26,8 +26,8 @@ export class IdeaPlugin extends StreamDeckPluginHandler {
         new DebugAction(this, "com.jetbrains.idea.debug");
         new EmptyAction(this, 'com.jetbrains.idea.empty.action');
         new GitPullAction(this, 'com.jetbrains.idea.git.pull');
-        new NewProjectAction(this, 'com.jetbrains.idea.new');
-        new OpenRecentAction(this, 'com.jetbrains.idea.open.recent');
+        new NewProjectAction(this, 'com.jetbrains.idea.project.new');
+        new OpenRecentAction(this, 'com.jetbrains.idea.project.open.recent');
         new PauseAction(this, 'com.jetbrains.idea.action.pause');
         new ResumeAction(this, 'com.jetbrains.idea.resume');
         new RunAction(this, 'com.jetbrains.idea.run');