+++ streamdeck-plugin/release_notes.md
@@ -1,4 +1,15 @@
+1.0.6
+
+- Order actions from `idea-plugin.ts` alphabetically
+- Move plugins to subfolders
+- Align naming convention for icons
+- Align size of icons
+- Fix "actions" icons for run and stop
+- Add `Open Recent` action
+- Add `Choose Config` action
+
 1.0.5
+
 - Change version from v1.0.4 to v1.0.5 (_`Step Into` action added without version change…_)
 - Change IntelliJ IDEA icon to use the 2024.2 version
 - Fix `Resume` button icon (_colored icon, same as other buttons_)